refactor: rename agent_foundation to agent and untrack conductor
