+++ README.md
@@ -39,7 +39,7 @@ uv sync
 ### 3) Run Locally
 
 ```bash
-uv run python -m agent_foundation.server
+uv run python -m agent.server
 ```
 Visit `http://127.0.0.1:8080`.
