+++ compose.yaml
@@ -18,7 +18,7 @@ services:
     volumes:
       - agent_artifacts:/app/src/.adk
     restart: always
-    command: python -m agent_foundation.server
+    command: python -m agent.server
 
 volumes:
   agent_artifacts:
