+++ docs/base-infra/dockerfile-strategy.md
@@ -340,7 +340,7 @@ EXPOSE 8080
 ### Startup Command
 ```dockerfile
 # Run the FastAPI server via main() for unified startup logic (logging, etc.)
-CMD ["python", "-m", "agent_foundation.server"]
+CMD ["python", "-m", "agent.server"]
 ```
 **What:** Default command when container starts
 **Why:**
