+++ systemd/agent.service
@@ -20,7 +20,7 @@ EnvironmentFile=/home/ubuntu/google-adk-on-bare-metal/.env
 
 # Path to 'uv' or 'python' executable
 # Assuming 'uv sync' has been run, we use the virtualenv python
-ExecStart=/home/ubuntu/google-adk-on-bare-metal/.venv/bin/python -m agent_foundation.server
+ExecStart=/home/ubuntu/google-adk-on-bare-metal/.venv/bin/python -m agent.server
 
 # Restart settings
 Restart=always
