+++ tests/test_logging_callbacks.py
@@ -17,7 +17,7 @@ from conftest import (
     MockToolContext,
 )
 
-from agent_foundation.callbacks import LoggingCallbacks
+from agent.callbacks import LoggingCallbacks
 
 # Note: Custom mock classes (conftest.py) use duck typing to match ADK interfaces.
 
@@ -39,7 +39,7 @@ class TestLoggerInjection:
         callbacks = LoggingCallbacks()
 
         assert callbacks.logger is not None
-        assert callbacks.logger.name == "agent_foundation.callbacks"
+        assert callbacks.logger.name == "agent.callbacks"
 
     def test_logging_callbacks_custom_logger(
         self, custom_logger: logging.Logger
